Add first attempt at button enabling

+++ src/Dispatch.java
@@ -2,6 +2,7 @@
  * File: Dispatch.java
  * Author: David Green DGreen@uab.edu
  * Assignment:  ScheduleBuilder - EE333 Fall 2018
+ * Vers: 1.1.0 11/24/2018 dgg - refactor, test file statuses
  * Vers: 1.0.0 11/19/2018 dgg - initial coding
  */
 
@@ -15,24 +16,35 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import static java.nio.file.StandardCopyOption.*;
 
-
 /**
  * Dispatch various external programs to perform desired behaviors
  * @author David Green DGreen@uab.edu
  */
 public class Dispatch {
     
+    String fullPath;
+    
     /**
-     * Dispatch the spreadsheet to allow schedule creation.Presently, don't hang 
-     * program but merely delay 10 seconds to allow the spreadsheet to show.
+     * Set the path the path for the data files
      * @param path path to courses
      * @param course specific course
      * @param semester specific semester
+     */
+    public void setFullPath(String path, String course, String semester) {
+        fullPath = path + "/" + course + "/" + semester;
+    }
+    
+    /**
+     * Dispatch the spreadsheet to allow schedule creation.Presently, don't hang 
+     * program but merely delay 10 seconds to allow the spreadsheet to show.
+     * Assumes fullPath has been set prior to execution.
      * @return String representation of result
      */
-    public String createEditSchedule(String path, String course, String semester) {
+    public String createEditSchedule() {
+        if (! readyForCreateEdit() ) {
+            return "";
+        }
         try {
-            String fullPath        = path + "/" + course + "/" + semester;
             String spreadSheetName = "schedule.numbers";        
             File schedule          = new File(fullPath + "/" + spreadSheetName);
             
@@ -56,15 +68,14 @@ public class Dispatch {
     
     /**
      * Edit the template behavior -- invoke editor assigned to .md suffix
-     * @param path path to courses
-     * @param course specific course
-     * @param semester specific semester
+     * Assumes fullPath has been set prior to execution.
      * @return String representation of result
      */
-    public String editTemplate(String path, String course, String semester) {
+    public String editTemplate() {
+        if (! readyForEditTemplate() ) {
+            return "";
+        }
         try {
-            String fullPath = path + "/" + course + "/" + semester;
-            
             Runtime rt = Runtime.getRuntime();
             Process p = rt.exec( "open schedule.md", (String[]) null, new File(fullPath) );
 
@@ -79,19 +90,19 @@ public class Dispatch {
     /**
      * Merge the CSV from the spreadsheet into the Template replacing any exising
      * old schedule
-     * @param path path to courses
-     * @param course specific course
-     * @param semester specific semester
+     * Assumes fullPath has been set prior to execution.
      * @return String representation of result     * @return
      */
-    public String mergeSchedule(String path, String course, String semester) {
+    public String mergeSchedule() {
         
+        if (! readyForMerge() ) {
+            return "";
+        }
+
         boolean finished = false;
 
         try {
-            String fullPath = path + "/" + course + "/" + semester;
-            
-            // ensure the schedule.old has been removed
+// ensure the schedule.old has been removed
             File file = new File(fullPath + "/schedule.old");
             file.delete();           // quietly continues even if file not there            
             
@@ -121,15 +132,14 @@ public class Dispatch {
     /**
      * Allow touching up file after merge (presently only beautification of .md)
      * using the editor assigned to .md suffix.
-     * @param path path to courses
-     * @param course specific course
-     * @param semester specific semester
+     * Assumes fullPath has been set prior to execution.
      * @return String representation of result
      */
-    public String touchUpResult(String path, String course, String semester) {
+    public String touchUpResult() {
+        if (! readyForMerge() ) {
+            return "";
+        }
         try {
-            String fullPath = path + "/" + course + "/" + semester;
-            
             Runtime rt = Runtime.getRuntime();
             Process p = rt.exec( "open schedule.md", (String[]) null, new File(fullPath) );
 
@@ -143,15 +153,14 @@ public class Dispatch {
     
     /**
      * View result in production view (using Marked which also reads the .md file)
-     * @param path path to courses
-     * @param course specific course
-     * @param semester specific semester
+     * Assumes fullPath has been set prior to execution.
      * @return String representation of result
      */
-    public String viewResult(String path, String course, String semester) {
-        try {
-            String fullPath = path + "/" + course + "/" + semester;
-            
+    public String viewResult() {
+        if (! readyForView() ) {
+            return "";
+        }
+        try {           
             Runtime rt = Runtime.getRuntime();
             Process p = rt.exec( "open schedule.md -a Marked.app", (String[]) null, new File(fullPath) );
 
@@ -165,18 +174,18 @@ public class Dispatch {
     
     /**
      * Export the schedule to HTML using multi-markdown script mmd
-     * @param path path to courses
-     * @param course specific course
-     * @param semester specific semester
+     * Assumes fullPath has been set prior to execution.
      * @return String representation of result
      */
-    public String exportResult(String path, String course, String semester) {
+    public String exportResult() {
+
+        if (! readyForExport() ) {
+            return "";
+        }
         
         boolean finished = false;
 
-        try {
-            String fullPath = path + "/" + course + "/" + semester;
-        
+        try {        
             ProcessBuilder pb = new ProcessBuilder("/Users/dgreen/bin/mmd", "schedule.md");
             pb.directory(new File(fullPath));
             File log = new File(fullPath + "/merge.log");
@@ -198,4 +207,90 @@ public class Dispatch {
         else
             return "Export complete";
     }
+    
+    /**
+     * Determine if Create/Edit action is allowable
+     * @return true if yes
+     */
+    public boolean readyForCreateEdit() {
+        return true;                        // always ready to do this
+    }
+    
+    /**
+     * Determine if Edit action is allowable
+     * @return true if yes
+     */
+    public boolean readyForEditTemplate() {
+        return scheduleMdExists();
+    }
+
+    /**
+     * Determine if Merge action is allowable
+     * @return true if yes
+     */
+    public boolean readyForMerge() {
+        return scheduleMdExists() && scheduleCSVExists();
+    }
+
+    /**
+     * Determine if Touch Up action is allowable
+     * @return true if yes
+     */
+    public boolean readyForTouchUp() {
+        return isScheduleMdNewerScheduleCSV();
+    }
+
+    /**
+     * Determine if View action is allowable
+     * @return true if yes
+     */
+    public boolean readyForView() {
+        return isScheduleMdNewerScheduleCSV();
+    }
+
+    /**
+     * Determine if Create/Edit action is allowable
+     * @return true if yes
+     */
+    public boolean readyForExport() {
+        return isScheduleMdNewerScheduleCSVScheduleHtml();
+    }
+    
+    private boolean scheduleMdExists() {
+        File scheduleMd = new File( fullPath + "/schedule.md");
+        return scheduleMd.exists();
+    }
+    
+    private boolean scheduleCSVExists() {
+        File scheduleCSV = new File( fullPath + "/schedule.csv");
+        return scheduleCSV.exists();
+    }
+
+    private boolean isScheduleMdNewerScheduleCSV() {
+        boolean result = false;
+        
+        File scheduleMd  = new File( fullPath + "/schedule.md");
+        File scheduleCSV = new File( fullPath + "/schedule.csv");
+        
+        if ( scheduleMd.exists() && scheduleCSV.exists() ) {
+            result = scheduleMd.lastModified() > scheduleCSV.lastModified();
+        }    
+        return result;   
+    }
+
+    private boolean isScheduleMdNewerScheduleCSVScheduleHtml() {
+        boolean result = false;
+        
+        File scheduleMd   = new File(fullPath + "/schedule.md");
+        File scheduleCSV  = new File(fullPath + "/schedule.csv");
+        File scheduleHtml = new File(fullPath + "/schedule.html");
+        
+        if ( scheduleMd.exists() && scheduleCSV.exists() ) {
+            if ( scheduleMd.lastModified() > scheduleCSV.lastModified() ) {
+                result = scheduleMd.lastModified() > scheduleHtml.lastModified();
+            }
+        }    
+        return result;   
+    }
+
 }
